Reworked UI tests to be more JavaFX-y, JUnit-y, sane
