+++ corona-ide-main/.corona-ide/com.coronaide.core/metadata
@@ -1,0 +1,1 @@
+{"projectLocations":[{"rootDirectory":"C:\\Users\\moldo\\Documents\\Eclipse\\corona-ide\\corona-ide\\corona-ide-main\\.corona-ide\\test"},{"rootDirectory":"C:\\Users\\moldo\\Documents\\Eclipse\\corona-ide\\corona-ide\\corona-ide-main\\.corona-ide\\bananas"},{"rootDirectory":"C:\\Users\\moldo\\Documents\\Eclipse\\corona-ide\\corona-ide\\corona-ide-main\\.corona-ide\\wookie"}]}
