+++ corona-ide-main/.corona-ide/com.coronaide.core/versions.json
@@ -1,0 +1,1 @@
+{"metadata":{"major":0,"minor":1,"micro":0}}
