+++ corona-ide-ui/src/main/java/com/coronaide/ui/CoronaUIApplication.java
@@ -2,7 +2,7 @@ package com.coronaide.ui;
 
 import java.util.Objects;
 
-import org.springframework.context.annotation.AnnotationConfigApplicationContext;
+import org.springframework.context.ApplicationContext;
 
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
@@ -15,17 +15,12 @@ public class CoronaUIApplication extends Application {
 
     private static FXMLLoader loader;
 
-    private static AnnotationConfigApplicationContext context;
+    private static ApplicationContext context;
 
-    /**
-     * Start the Corona UI.
-     * 
-     * @param springContext
-     *            The Spring application context
-     */
-    public static void launchUi(AnnotationConfigApplicationContext springContext) {
+    private static Stage primaryStage;
+
+    public static void setSpringContext(ApplicationContext springContext) {
         context = Objects.requireNonNull(springContext);
-        Application.launch(CoronaUIApplication.class);
     }
 
     @Override
@@ -37,12 +32,25 @@ public class CoronaUIApplication extends Application {
 
     @Override
     public void start(Stage stage) throws Exception {
+        primaryStage = stage;
         Parent root = loader.load();
         Scene scene = new Scene(root, 1000, 500);
-        stage.setTitle("Corona");
         stage.setScene(scene);
         stage.getIcons().add(new Image(getClass().getResourceAsStream("/icon.png")));
         stage.show();
+    }
+
+    /**
+     * @return The application's primary stage, so that operations can be performed on it such as changing the title
+     * @throws IllegalStateException
+     *             if the primary stage has not yet been set
+     * @since 0.1.0
+     */
+    public static Stage getPrimaryStage() throws IllegalStateException {
+        if (primaryStage == null) {
+            throw new IllegalStateException("The primary Stage has not yet been set");
+        }
+        return primaryStage;
     }
 
 }
