+++ corona-ide-ui/src/main/resources/fxml/Main.fxml
@@ -1,6 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
-<?import javafx.geometry.Insets?>
 <?import javafx.scene.control.Label?>
 <?import javafx.scene.control.ListView?>
 <?import javafx.scene.control.Menu?>
@@ -10,7 +9,6 @@
 <?import javafx.scene.layout.BorderPane?>
 <?import javafx.scene.layout.HBox?>
 <?import javafx.scene.layout.Pane?>
-<?import javafx.scene.layout.VBox?>
 
 <BorderPane xmlns="http://javafx.com/javafx/8.0.111" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.coronaide.ui.controller.MainController">
    <top>
@@ -44,18 +42,6 @@
       </Pane>
    </center>
    <left>
-      <VBox prefHeight="200.0" prefWidth="192.0" BorderPane.alignment="CENTER">
-         <children>
-            <Label fx:id="labelWorkspace" text="workspace" wrapText="true">
-               <padding>
-                  <Insets bottom="5.0" left="10.0" right="10.0" top="5.0" />
-               </padding>
-               <VBox.margin>
-                  <Insets />
-               </VBox.margin>
-            </Label>
-            <ListView fx:id="listViewProjects" prefHeight="200.0" prefWidth="120.0" VBox.vgrow="ALWAYS" />
-         </children>
-      </VBox>
+      <ListView fx:id="listViewProjects" prefHeight="200.0" prefWidth="150.0" BorderPane.alignment="CENTER" />
    </left>
 </BorderPane>
